Move Streaming/Faye extensions into new `StreamingExtension` module
