+++ lib/core.js
@@ -14,3 +14,4 @@ jsforce.Date = jsforce.SfDate = require("./date");
 jsforce.RecordStream = require('./record-stream');
 jsforce.Promise = require('./promise');
 jsforce.require = require('./require');
+jsforce.StreamingExtension = require('./streaming-extension');
